Start on Simulation
Rewrote the game loop so that each part of the program (ie menu, simulation, quiz) uses its own loop. This way the buttons from the menu can easily be used to call the different parts of the program.
Also started on some simulation logic, namely moving a circle from point a to b. (This will be for the photons to move towards the plate, and the photoelectrons being emitted)